% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/agg.R
\name{aggregation_factory}
\alias{aggregation_factory}
\title{Aggregation function factory}
\usage{
aggregation_factory(
  specific_function,
  x,
  p = 0.1,
  q = 0.05,
  drop_zeroes = FALSE
)
}
\arguments{
\item{specific_function}{The specific aggregation function to be generated}

\item{x}{A vector of forecasts}

\item{p}{The proportion of forecasts to trim from each end (between 0 and
1)}

\item{q}{The quantile to use for replacing 0s and 1s (between 0 and 1)}
}
\description{
This is sort of a base class or a template for the individual
aggregation functions. It does the things that they all have in common.
}
