% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graphing.R
\name{boxPlot}
\alias{boxPlot}
\title{Basic boxplot function}
\usage{
boxPlot(
  files,
  type,
  specialty,
  title,
  subtitle,
  filenameStart,
  expectedRisk,
  forecastMin,
  forecastMax,
  numerateCitizens,
  yLabel,
  setName,
  beliefSet,
  year,
  distrib
)
}
\description{
Basic boxplot function
}
