% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graphing.R
\name{boxPlot_distrib}
\alias{boxPlot_distrib}
\title{Box Plot for Distribution Questions}
\usage{
boxPlot_distrib(
  tbl,
  specialty,
  title,
  forecastMin,
  forecastMax,
  stage,
  year,
  numerateCitizens,
  yLabel,
  setName,
  beliefSet,
  distrib
)
}
\description{
Box Plot for Distribution Questions
}
