% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graphing.R
\name{boxPlot_distrib_country}
\alias{boxPlot_distrib_country}
\title{Box Plot for Distribution Country Questions}
\usage{
boxPlot_distrib_country(
  tbl,
  specialty,
  title,
  forecastMin,
  forecastMax,
  stage,
  year
)
}
\description{
Box Plot for Distribution Country Questions
}
