% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convergence.R
\name{convergenceAdd_active}
\alias{convergenceAdd_active}
\title{Subset to forecasters who were really committed, still active in Phase 4.}
\usage{
convergenceAdd_active(phase1, phase2, phase3, phase4, convergenceTable)
}
\arguments{
\item{phase1}{TODO}

\item{phase2}{TODO}

\item{phase3}{TODO}

\item{phase4}{TODO}

\item{convergenceTable}{TODO}
}
\description{
Subset to forecasters who were really committed, still active in Phase 4.
}
