% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convergence.R
\name{convergenceInit}
\alias{convergenceInit}
\title{Initialize empty convergence table.}
\usage{
convergenceInit()
}
\value{
Empty table with convergence columns.
}
\description{
How much did teams converge (within-team)?
}
