% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{getPublicSurveyForecasts}
\alias{getPublicSurveyForecasts}
\title{Get public survey forecasts for a given question.}
\usage{
getPublicSurveyForecasts(sn, bs, y, d)
}
\arguments{
\item{sn}{\itemize{
\item set name
}}

\item{bs}{\itemize{
\item belief set
}}

\item{y}{\itemize{
\item year
}}

\item{d}{\itemize{
\item distribution
}}
}
\description{
This is only called by make_agg_methods_table() and is not exported.
}
