% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/agg.R
\name{hd_trim}
\alias{hd_trim}
\title{Highest-Density Trimmed Mean}
\usage{
hd_trim(x, p = 0.1)
}
\arguments{
\item{x}{A vector of forecasts}

\item{p}{The proportion of forecasts to trim (between 0 and 1)}
}
\description{
This code comes from an email from Ben Powell.

You find the shortest interval containing (1-p) * 100\% of the data and take the
mean of the forecasts within that interval.
}
\note{
As p gets bigger this acts like a mode in a similar way to
the symmetrically-trimmed mean acting like a median.
}
