% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graphing.R
\name{histogram}
\alias{histogram}
\title{Histogram}
\usage{
histogram(
  questionDataProcessed,
  filenameStart,
  title,
  stage,
  specialty,
  expectedRisk,
  forecastMin,
  forecastMax
)
}
\description{
Histogram
}
