% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{make_agg_methods_table}
\alias{make_agg_methods_table}
\title{Make a table for a given question with a few different aggregation
methods, by group. Groups and methods currently hardcoded.}
\usage{
make_agg_methods_table("AI Catastrophic Risk", 2100)
}
\description{
Make a table for a given question with a few different aggregation
methods, by group. Groups and methods currently hardcoded.
}
