% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graphing.R
\name{multiCountryBinaryGraphics}
\alias{multiCountryBinaryGraphics}
\title{Multi-Country Binary Graphics}
\usage{
multiCountryBinaryGraphics(title, subtitle, csv, currentSetName, country)
}
\description{
Multi-Country Binary Graphics
}
