% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{multiYearBinary}
\alias{multiYearBinary}
\title{Stats and graphs for Multi-year Binary Questions}
\usage{
multiYearBinary(metaTable, data)
}
\arguments{
\item{metaTable}{\itemize{
\item metadata for all the multi-year reciprocal question sets
}}

\item{data}{\itemize{
\item data for all users on all multi-year reciprocal question sets
}}
}
\description{
For yes/no multi-year questions, like:
What is the probability that the use of a nuclear weapon (in a single
event) will cause the death of more than 1,000 people...
...by 2024?
...by 2030?

Builds out the folder structure, populates with stats and graphs.
}
