% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rs.R
\name{multiYearBinary_RS}
\alias{multiYearBinary_RS}
\title{Scores the multi-year binary questions}
\usage{
multiYearBinary_RS(metaTable, data, summaryTable)
}
\arguments{
\item{metaTable}{Metadata on the multi-year binary reciprocal questions}

\item{data}{Data on the multi-year reciprocal binary questions (forecasts)}

\item{summaryTable}{TODO}
}
\description{
This scores everyone who answered the multi-year binary
reciprocal questions (What do you believe? Ok, what do you think the \link[=other
supers, experts]{other supers, experts} believe?). They were asked about the MEDIAN belief of two
groups, for each year (yes/no).

Scoring rule:
\verb{log(1 - abs(action median - what you thought the median would be))}
}
\note{
"Median" belief on yes/no questions seems silly; did we ask any of these
}
