% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graphing.R
\name{multiYearBinaryFigureData}
\alias{multiYearBinaryFigureData}
\title{Multi-year Binary Figure Data}
\usage{
multiYearBinaryFigureData(metaTable, data, phaseTwoMedian, timeline)
}
\description{
Multi-year Binary Figure Data
}
