% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graphing.R
\name{multiYearBinaryGraphics}
\alias{multiYearBinaryGraphics}
\title{Multi-year Binary Graphics}
\usage{
multiYearBinaryGraphics(title, subtitle, csv, currentSetName, year)
}
\description{
Multi-year Binary Graphics
}
