% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graphing.R
\name{multiYearBinaryVarianceGraphics}
\alias{multiYearBinaryVarianceGraphics}
\title{Multi-year Binary Graphics}
\usage{
multiYearBinaryVarianceGraphics(title, subtitle, csv, currentSetName, year)
}
\description{
Multi-year Binary Graphics
}
