% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{multiYearCountryDistrib}
\alias{multiYearCountryDistrib}
\title{Stats and graphs for Multi-year Country Distribution Questions}
\usage{
multiYearCountryDistrib(metaTable, data)
}
\arguments{
\item{metaTable}{\itemize{
\item metadata for all the multi-year country distribution question sets
}}

\item{data}{\itemize{
\item data for all users on all multi-year country distribution question sets
}}
}
\description{
For questions where we ask for distributions over predefined
years AND a set of countries, this function will build out the folder structure
and populate it with stats and graphs.
}
