% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rs.R
\name{multiYearCountryDistrib_RS}
\alias{multiYearCountryDistrib_RS}
\title{Scores the multi-year distribution reciprocal question sets where country also varied}
\usage{
multiYearCountryDistrib_RS(metaTable, data, summaryTable)
}
\arguments{
\item{metaTable}{Metadata on the multi-year country distribution reciprocal questions}

\item{data}{Data on the multi-year country distribution reciprocal questions (forecasts)}

\item{summaryTable}{TODO}
}
\description{
This scores everyone who answered the reciprocal questions
(What do you believe? Ok, what do you think the \link{other supers, experts}
believe?). They were asked about the MEDIAN belief of two groups, for each
year and country combination.

Scoring rule:
\verb{log(1 - abs(action median - what you thought the median would be))}
}
\note{
Do we use this
}
