% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graphing.R
\name{multiYearCountryVarianceGraphics}
\alias{multiYearCountryVarianceGraphics}
\title{Multi-Year Country Variance Graphics}
\usage{
multiYearCountryVarianceGraphics(
  title,
  subtitle,
  csv,
  currentSetName,
  year,
  country
)
}
\description{
Multi-Year Country Variance Graphics
}
