% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{multiYearDistrib}
\alias{multiYearDistrib}
\title{Stats and graphs for Multi-year Distribution Questions}
\usage{
multiYearDistrib(
  metaTable,
  data,
  public_supplement1,
  public_supplement2,
  public_supplement3,
  survey_column_matches
)
}
\arguments{
\item{metaTable}{\itemize{
\item metadata for all the multi-year distribution question sets
}}

\item{data}{\itemize{
\item data for all users on all multi-year distribution question sets
}}
}
\description{
For the questions where we ask for distributions over predefined years, like:
How many total nuclear warheads will be in military inventories globally by
...the end of 2024?
...the end of 2030?
...the end of 2040?

Builds out the folder structure, populates with stats and graphs.
}
