% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rs.R
\name{multiYearDistrib_RS}
\alias{multiYearDistrib_RS}
\title{Scores the multi-year distribution reciprocal questions}
\usage{
multiYearDistrib_RS(metaTable, data, summaryTable)
}
\arguments{
\item{metaTable}{Metadata on the multi-year distribution reciprocal questions}

\item{data}{Data on the multi-year distribution reciprocal questions (forecasts)}

\item{summaryTable}{TODO}
}
\description{
This scores everyone who answered the multi-year distribution
reciprocal questions (What do you believe? Ok, what do you think the \link[=other
supers, experts]{other supers, experts} believe?). They were asked about the MEDIAN belief of two
groups, for each year and quintile combination.

Scoring rule:
\verb{log(1 - abs(action median - what you thought the median would be))}
}
\note{
Were there any of these questions?
}
