% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graphing.R
\name{multiYearDistribFigureData}
\alias{multiYearDistribFigureData}
\title{Multi-year Distribution Figure Data}
\usage{
multiYearDistribFigureData(metaTable, data, phaseTwoMedian, timeline)
}
\description{
Multi-year Distribution Figure Data
}
