% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graphing.R
\name{multiYearDistribTeamGraphics}
\alias{multiYearDistribTeamGraphics}
\title{Multi-Year Distribution Questions: Team Graphics}
\usage{
multiYearDistribTeamGraphics(
  title,
  subtitle,
  csv,
  currentSetName,
  distrib,
  year
)
}
\description{
All members of all teams
}
