% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{multiYearReciprocal}
\alias{multiYearReciprocal}
\title{Stats and graphs for the multi-year questions}
\usage{
multiYearReciprocal(
  metaTable,
  data,
  main1,
  main2,
  supplement,
  survey_column_matches
)
}
\arguments{
\item{metaTable}{\itemize{
\item metadata for all the multi-year reciprocal question sets
}}

\item{data}{\itemize{
\item data for all users on all multi-year reciprocal question sets
}}
}
\description{
For the question sets that asked about a forecaster's own
beliefs and what they thought the supers and the experts believed.
This function builds out a folder structure and populates it with stats and graphs for these questions.
}
\note{
All reciprocal questions were multi-year.
}
