% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rs.R
\name{multiYearReciprocal_RS}
\alias{multiYearReciprocal_RS}
\title{Scores the multi-year reciprocal questions}
\usage{
multiYearReciprocal_RS(metaTable, data, summaryTable)
}
\arguments{
\item{metaTable}{Metadata on the multi-year reciprocal questions}

\item{data}{Data on the multi-year reciprocal questions (forecasts)}

\item{summaryTable}{TODO}
}
\description{
This scores everyone who answered the reciprocal questions
(What do you believe? Ok, what do you think the \link{other supers, experts}
believe?). They were asked about the MEDIAN belief of two groups.

Scoring rule:
\verb{log(1 - abs(actual median - what you thought the median would be))}

+------------+-----------------------+-------------------------------------+ # nolint
|     x      |       A. Actual       | B. What _ believes Group X will say | # nolint
+------------+-----------------------+-------------------------------------+ # nolint
| 1. Yours   | Your true belief      | Your belief about Group X           | # nolint
| 2. Group X | True Group X (median) | What Group X believes about itself  | # nolint
+------------+-----------------------+-------------------------------------+ # nolint

This produces two kinds of scores:
score_unincentivized compares A2 to B1, and
score_rs compares B2 to B1
}
