% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graphing.R
\name{multiYearReciprocalFigureData}
\alias{multiYearReciprocalFigureData}
\title{Multi-year Reciprocal Figure Data}
\usage{
multiYearReciprocalFigureData(metaTable, data, phaseTwoMedian, timeline)
}
\description{
Multi-year Reciprocal Figure Data
}
