% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{newRowInit}
\alias{newRowInit}
\title{TODO Molly :)}
\usage{
newRowInit(
  metaTable,
  questionDataProcessed,
  currentSetName,
  currentQuestionName,
  answerText,
  stage,
  specialty
)
}
\description{
TODO Molly :)
}
