% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/agg.R
\name{neymanAggCalc}
\alias{neymanAggCalc}
\title{Neyman Aggregation (Extremized)}
\usage{
neymanAggCalc(x, p, q, drop_zeroes)
}
\arguments{
\item{x}{A vector of forecasts}
}
\description{
Origin: Neyman and Roughgarden 2021
See also Jaime Savilla's post on the EA Forum about it.

Extremizes the aggregated forecast by a factor of:
(n*(sqrt((3\emph{n^2) - (3}n) + 1) - 2))/(n^2 - n - 1)
where n is the number of forecasts.
}
\note{
Expects forecasts to be in the range \link{0, 100}!
}
\references{
Neyman, E. and Roughgarden, T. (2021). Are you
smarter than a random expert? The robust aggregation of substitutable
signals. \verb{https://arxiv.org/abs/2111.03153}
}
