% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graphing.R
\name{plot_with_ribbons}
\alias{plot_with_ribbons}
\title{Plot time series with confidence interval ribbons.}
\usage{
plot_with_ribbons(plotTable, title, subtitle, phaseTwoMedian)
}
\arguments{
\item{plotTable}{Data frame with columns currentDate, median, group}

\item{title}{Title of plot}

\item{subtitle}{Subtitle of plot}

\item{phaseTwoMedian}{Date of median start time for Phase 2 (we cut the plot off on the left
at this date)}
}
\description{
Plot time series with confidence interval ribbons.
}
