% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/629.R
\name{pointBinary_629}
\alias{pointBinary_629}
\title{Stats and graphs for classic binary questions}
\usage{
pointBinary_629(
  metaTable,
  data,
  public_supplement1,
  public_supplement2,
  public_supplement3,
  survey_column_matches
)
}
\arguments{
\item{metaTable}{\itemize{
\item metadata for all the point binary question sets
}}

\item{data}{\itemize{
\item data for all users on all point binary question sets
}}
}
\description{
For yes/no point questions, like
"Will Robin Hanson win a bet that the GPT line of language models will
generate less than $1 billion in customer revenue in total by the beginning
of 2025?"

This function builds out a folder structure and populates it with stats and
graphs.
}
