% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graphing.R
\name{pointBinaryVarianceGraphics}
\alias{pointBinaryVarianceGraphics}
\title{Point Binary Variance Graphics}
\usage{
pointBinaryVarianceGraphics(title, subtitle, csv, currentSetName)
}
\description{
Point Binary Variance Graphics
}
