% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/629.R
\name{pointDistrib_629}
\alias{pointDistrib_629}
\title{Stats and Graphs for Point Distribution Questions}
\usage{
pointDistrib_629(
  metaTable,
  data,
  public_supplement1,
  public_supplement2,
  public_supplement3,
  survey_column_matches
)
}
\arguments{
\item{metaTable}{\itemize{
\item metadata for all the point distribution question sets
}}

\item{data}{\itemize{
\item data for all users on all point distribution question sets
}}
}
\description{
For questions where we just asked for a distribution, like:
By what year will humans go extinct or first have a population less than 5,000?

Builds out the folder structure, populates with stats and graphs.
}
