% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rs.R
\name{pointDistrib_RS}
\alias{pointDistrib_RS}
\title{Scores the point distribution reciprocal questions}
\usage{
pointDistrib_RS(metaTable, data, summaryTable)
}
\arguments{
\item{metaTable}{Metadata on the point distrib reciprocal questions}

\item{data}{Data on the point distrib reciprocal questions (forecasts)}

\item{summaryTable}{TODO}
}
\description{
This scores everyone who answered the reciprocal questions
(What do you believe? Ok, what do you think the \link{other supers, experts}
believe?). They were asked about the MEDIAN belief of two groups, for
each quintile.

Scoring rule:
\verb{log(1 - abs(action median - what you thought the median would be))}
}
\note{
Were there any PD RS questions?
}
