% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graphing.R
\name{pointDistribFigureData}
\alias{pointDistribFigureData}
\title{Point Distribution Figure Data}
\usage{
pointDistribFigureData(metaTable, data, phaseTwoMedian, timeline)
}
\description{
Point Distribution Figure Data
}
