% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/agg.R
\name{preprocess}
\alias{preprocess}
\title{Preprocessing function for agg methods}
\usage{
preprocess(x, q = 0)
}
\arguments{
\item{x}{A vector of forecasts}

\item{q}{The quantile to use for replacing 0s and 1s (between 0 and 1)}

\item{specific_function}{The specific aggregation function to be generated}
}
\description{
This does the preprocessing steps that all the agg methods
have in common.
}
\note{
ASSUMES FORECASTS ARE IN THE RANGE \link{0, 100}!
}
