% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rs.R
\name{RSInit}
\alias{RSInit}
\title{Initialize reciprocal scoring dataframe.}
\usage{
RSInit()
}
\value{
An empty dataframe with the reciprocal scoring columns.
}
\description{
Initialize reciprocal scoring dataframe.
}
