% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rs.R
\name{RSRankingInit}
\alias{RSRankingInit}
\title{Initialize reciprocal scoring ranking dataframe.}
\usage{
RSRankingInit()
}
\value{
An empty dataframe that will hold the reciprocal scoring ranking
for each user.
}
\description{
Initialize reciprocal scoring ranking dataframe.
}
