% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/salience.R
\name{salienceCalc}
\alias{salienceCalc}
\title{Calculate salience.}
\usage{
salienceCalc(
  median,
  sd,
  g1Median,
  g1Sd,
  supersMedian,
  dExpertsMedian,
  ndExpertsMedian,
  csv,
  expPosSalience
)
}
\arguments{
\item{median}{TODO}

\item{sd}{\itemize{
\item 
}}

\item{g1Median}{\itemize{
\item 
}}

\item{g1Sd}{\itemize{
\item 
}}

\item{supersMedian}{\itemize{
\item 
}}

\item{dExpertsMedian}{\itemize{
\item 
}}

\item{ndExpertsMedian}{\itemize{
\item 
}}

\item{csv}{\itemize{
\item 
}}

\item{expPosSalience}{\itemize{
\item 
}}
}
\description{
We wanted to track, over time, how much more optimistic or pessimistic
supers and (g1) experts were compared to the median of the two groups
combined after the independent phase of forecasting.
}
