% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/agg.R
\name{soften_mean}
\alias{soften_mean}
\title{Soften the mean.}
\usage{
soften_mean(x, p = 0.1, drop_zeroes = FALSE)
}
\arguments{
\item{x}{A vector of forecasts}

\item{drop_zeroes}{Whether to drop zeroes from the vector}

\item{trim}{The proportion of forecasts to trim from each end (between 0
and 1)}
}
\description{
If the mean is > .5, trim the top trim\%; if < .5, the bottom trim\%. Return
the new mean (i.e. soften the mean).
}
\note{
This goes against usual wisdom of extremizing the mean, but performs
well when the crowd has some overconfident forecasters in it.
}
