% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graphing.R
\name{boxPlot_country}
\alias{boxPlot_country}
\title{Box Plot for Country Questions}
\usage{
boxPlot_country(
  tbl,
  specialty,
  title,
  forecastMin,
  forecastMax,
  stage,
  numerateCitizens,
  yLabel,
  setName,
  beliefSet,
  year,
  distrib
)
}
\description{
Box Plot for Country Questions
}
