% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/agg.R
\name{trim}
\alias{trim}
\title{Trimmed mean}
\usage{
trim(x, p = 0.1)
}
\arguments{
\item{p}{The proportion of forecasts to trim from each end (between 0 and 1)}

\item{questionData}{The processed question data table (needs to have a forecast column)}
}
\value{
The trimmed mean
}
\description{
Trim the top and bottom (p*100)\% of forecasts
}
